backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
sqlalchemy==2.0.32
supabase==2.5.1
psycopg[binary]==3.2.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
reportlab==4.2.2
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.23.8
httpx==0.27.2
